requirements.txt
flask
numpy
scikit-learn
Pillow
torch
torchvision
open-clip-torch
pandas
